ai/aisignal/requirements.txt
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.4
deprecation==2.1.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
google-generativeai==0.8.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.1
narwhals==2.16.0
networkx==3.6.1
numpy==2.4.2
packaging==26.0
pandas==2.3.3
pillow==12.1.0
postgrest==2.27.3
propcache==0.4.1
propcache==0.4.1
protobuf<6.0.0
psycopg2-binary==2.9.11
pyarrow==23.0.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.11.0
pyparsing==3.3.2
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
pyvis==0.3.2
realtime==2.27.3
redis==7.1.1
referencing==0.37.0
requests==2.32.5
rich==14.3.2
rpds-py==0.30.0
six==1.17.0
smmap==5.0.2
sortedcontainers==2.4.0
storage3==2.27.3
streamlit==1.54.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.27.3
supabase-auth==2.27.3
supabase-functions==2.27.3
tenacity==9.1.4
toml==0.10.2
tornado==6.5.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
websockets==15.0.1
yarl==1.22.0
yfinance==0.2.36
fastapi==0.109.2
uvicorn==0.27.1
